deleted Setting folder and moved changeDorm
